Finished migration to uniffi

+++ domain/src/main/kotlin/com/jetpackduba/gitnuro/domain/libssh/streams/SshChannelInputErrStream.kt
@@ -1,6 +1,6 @@
 package com.jetpackduba.gitnuro.domain.libssh.streams
 
-import Channel
+import com.jetpackduba.gitnuro.Channel
 import com.jetpackduba.gitnuro.domain.exceptions.SshException
 import java.io.InputStream
 
@@ -9,7 +9,7 @@ class SshChannelInputErrStream(private val sshChannel: Channel) : InputStream() 
 
     override fun read(): Int {
         return if (sshChannel.pollHasBytes(true)) {
-            val read = sshChannel.read(true, 1L)
+            val read = sshChannel.read(true, 1L.toULong())
                 ?: throw SshException("Could not read result from SSH channel. Please check your network connectivity and try again.")
 
             val byteArray = read.data
